chore: extract the entire function that generates the query to the client
