+++ packages/query/src/client.ts
@@ -6,19 +6,22 @@ import {
   GeneratorMutator,
   GetterResponse,
   pascal,
-  GetterProps,
-  ContextSpecs,
   GeneratorDependency,
   GeneratorOptions,
   OutputHttpClient,
+  VERBS_WITH_BODY,
+  generateFormDataAndUrlEncodedFunction,
+  generateMutatorConfig,
+  generateMutatorRequestOptions,
 } from '@orval/core';
 
-import {
-  generateRequestFunction as generateFetchRequestFunction,
-  fetchResponseTypeName,
-} from '@orval/fetch';
+import { generateRequestFunction as generateFetchRequestFunction } from '@orval/fetch';
 
-import { vueUnRefParams } from './utils';
+import {
+  makeRouteSafe,
+  vueUnRefParams,
+  vueWrapTypeWithMaybeRef,
+} from './utils';
 
 export const AXIOS_DEPENDENCIES: GeneratorDependency[] = [
   {
@@ -37,48 +40,13 @@ export const AXIOS_DEPENDENCIES: GeneratorDependency[] = [
   },
 ];
 
-export const generateRequestOptionsArguments = ({
-  isRequestOptions,
-  hasSignal,
-}: {
-  isRequestOptions: boolean;
-  hasSignal: boolean;
-}) => {
-  if (isRequestOptions) {
-    return 'options?: AxiosRequestConfig\n';
-  }
-
-  return hasSignal ? 'signal?: AbortSignal\n' : '';
-};
-
-export const generateQueryHttpRequestFunction = (
+export const generateQueryRequestFunction = (
   verbOptions: GeneratorVerbOptions,
-  context: ContextSpecs,
   options: GeneratorOptions,
-  props: GetterProps,
-  route: string,
   isVue: boolean,
-  isRequestOptions: boolean,
-  isFormData: boolean,
-  isFormUrlEncoded: boolean,
-  hasSignal: boolean,
-  isExactOptionalPropertyTypes: boolean,
-  bodyForm: string,
 ) => {
   if (options.context.output.httpClient === OutputHttpClient.AXIOS) {
-    return generateAxiosRequestFunction(
-      verbOptions,
-      context,
-      props,
-      route,
-      isVue,
-      isRequestOptions,
-      isFormData,
-      isFormUrlEncoded,
-      hasSignal,
-      isExactOptionalPropertyTypes,
-      bodyForm,
-    );
+    return generateAxiosRequestFunction(verbOptions, options, isVue);
   } else {
     return generateFetchRequestFunction(verbOptions, options);
   }
@@ -90,22 +58,115 @@ export const generateAxiosRequestFunction = (
     queryParams,
     operationName,
     response,
+    mutator,
     body,
+    props: _props,
     verb,
-    paramsSerializer,
+    formData,
+    formUrlEncoded,
     override,
+    paramsSerializer,
   }: GeneratorVerbOptions,
-  context: ContextSpecs,
-  props: GetterProps,
-  route: string,
+  { route: _route, context }: GeneratorOptions,
   isVue: boolean,
-  isRequestOptions: boolean,
-  isFormData: boolean,
-  isFormUrlEncoded: boolean,
-  hasSignal: boolean,
-  isExactOptionalPropertyTypes: boolean,
-  bodyForm: string,
 ) => {
+  let props = _props;
+  let route = _route;
+
+  if (isVue) {
+    props = vueWrapTypeWithMaybeRef(_props);
+  }
+
+  if (context.output?.urlEncodeParameters) {
+    route = makeRouteSafe(route);
+  }
+
+  const isRequestOptions = override.requestOptions !== false;
+  const isFormData = override.formData !== false;
+  const isFormUrlEncoded = override.formUrlEncoded !== false;
+  const hasSignal = !!override.query.signal;
+
+  const isExactOptionalPropertyTypes =
+    !!context.output.tsconfig?.compilerOptions?.exactOptionalPropertyTypes;
+  const isBodyVerb = VERBS_WITH_BODY.includes(verb);
+
+  const bodyForm = generateFormDataAndUrlEncodedFunction({
+    formData,
+    formUrlEncoded,
+    body,
+    isFormData,
+    isFormUrlEncoded,
+  });
+
+  if (mutator) {
+    const mutatorConfig = generateMutatorConfig({
+      route,
+      body,
+      headers,
+      queryParams,
+      response,
+      verb,
+      isFormData,
+      isFormUrlEncoded,
+      isBodyVerb,
+      hasSignal,
+      isExactOptionalPropertyTypes,
+      isVue,
+    });
+
+    let bodyDefinition = body.definition.replace('[]', '\\[\\]');
+    let propsImplementation =
+      mutator?.bodyTypeName && body.definition
+        ? toObjectString(props, 'implementation').replace(
+            new RegExp(`(\\w*):\\s?${bodyDefinition}`),
+            `$1: ${mutator.bodyTypeName}<${body.definition}>`,
+          )
+        : toObjectString(props, 'implementation');
+
+    const requestOptions = isRequestOptions
+      ? generateMutatorRequestOptions(
+          override.requestOptions,
+          mutator.hasSecondArg,
+        )
+      : '';
+
+    if (mutator.isHook) {
+      return `${
+        override.query.shouldExportMutatorHooks ? 'export ' : ''
+      }const use${pascal(operationName)}Hook = () => {
+        const ${operationName} = ${mutator.name}<${
+          response.definition.success || 'unknown'
+        }>();
+
+        return useCallback((\n    ${propsImplementation}\n ${
+          isRequestOptions && mutator.hasSecondArg
+            ? `options${context.output.optionsParamRequired ? '' : '?'}: SecondParameter<ReturnType<typeof ${mutator.name}>>,`
+            : ''
+        }${
+          !isBodyVerb && hasSignal ? 'signal?: AbortSignal\n' : ''
+        }) => {${bodyForm}
+        return ${operationName}(
+          ${mutatorConfig},
+          ${requestOptions});
+        }, [${operationName}])
+      }
+    `;
+    }
+
+    return `${override.query.shouldExportHttpClient ? 'export ' : ''}const ${operationName} = (\n    ${propsImplementation}\n ${
+      isRequestOptions && mutator.hasSecondArg
+        ? `options${context.output.optionsParamRequired ? '' : '?'}: SecondParameter<typeof ${mutator.name}>,`
+        : ''
+    }${!isBodyVerb && hasSignal ? 'signal?: AbortSignal\n' : ''}) => {
+      ${isVue ? vueUnRefParams(props) : ''}
+      ${bodyForm}
+      return ${mutator.name}<${response.definition.success || 'unknown'}>(
+      ${mutatorConfig},
+      ${requestOptions});
+    }
+  `;
+  }
+
   const isSyntheticDefaultImportsAllowed = isSyntheticDefaultImportsAllow(
     context.output.tsconfig,
   );
@@ -134,7 +195,7 @@ export const generateAxiosRequestFunction = (
 
   const queryProps = toObjectString(props, 'implementation');
 
-  return `export const ${operationName} = (\n    ${queryProps} ${optionsArgs} ): Promise<AxiosResponse<${
+  const httpRequestFunctionImplementation = `export const ${operationName} = (\n    ${queryProps} ${optionsArgs} ): Promise<AxiosResponse<${
     response.definition.success || 'unknown'
   }>> => {${bodyForm}
     ${isVue ? vueUnRefParams(props) : ''}
@@ -143,6 +204,22 @@ export const generateAxiosRequestFunction = (
     }.${verb}(${options});
   }
 `;
+
+  return httpRequestFunctionImplementation;
+};
+
+export const generateRequestOptionsArguments = ({
+  isRequestOptions,
+  hasSignal,
+}: {
+  isRequestOptions: boolean;
+  hasSignal: boolean;
+}) => {
+  if (isRequestOptions) {
+    return 'options?: AxiosRequestConfig\n';
+  }
+
+  return hasSignal ? 'signal?: AbortSignal\n' : '';
 };
 
 export const getQueryArgumentsRequestType = (
